+++ include/boost/http_proto.hpp
@@ -16,6 +16,7 @@
 #include <boost/http_proto/error.hpp>
 #include <boost/http_proto/field.hpp>
 #include <boost/http_proto/fields.hpp>
+#include <boost/http_proto/fields_base.hpp>
 #include <boost/http_proto/fields_view.hpp>
 #include <boost/http_proto/fields_view_base.hpp>
 #include <boost/http_proto/header_info.hpp>
